requirements.txt
aiogram==3.13.1
supabase==2.9.1
python-decouple==3.8
openpyxl==3.1.5
pandas==2.2.2
gunicorn==23.0.0
aiohttp==3.10.5
